gelab-zero-main/requirements.txt
megfile
opencv-python
jsonlines
pydantic
fastapi
uvicorn

pymongo
openai==0.28.0

pandas
Pillow
openpyxl

streamlit
tqdm
requests